Rename GoCard card folder to `card`
Forms for inputting login information will be part of a provider so the
generic 'component' isn't descriptive enough anymore. This clarifies
what that folder is for and makes it easier to understand.
