+++ providers/go_card/card/images/go.png
