+++ src/components/card_list/component.js
@@ -1,6 +1,5 @@
 import React from 'react'
 
-import GoCard from '../../../providers/go_card/component'
 import RefreshableList from '../refreshable_list'
 
 const CardList = ({ items, loading, onRefresh }) => (
